pyproject.toml
[project]
name = "giga-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref>=3.9.1",
    "bcrypt>=4.3.0",
    "django>=5.2.6",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "whitenoise>=6.9.0",
]
